refactor: collect RNGs under data.rng module
