+++ src/tensorwaves/interface.py
@@ -206,7 +206,10 @@ class Optimizer(ABC):
 
 
 class RealNumberGenerator(ABC):
-    """Abstract class for generating real numbers within a certain range."""
+    """Abstract class for generating real numbers within a certain range.
+
+    Implementations can be found in the `tensorwaves.data.rng` module.
+    """
 
     @abstractmethod
     def __call__(
