+++ tests/data/test_rng.py
@@ -1,0 +1,33 @@
+# pylint:disable=no-self-use
+import pytest
+
+from tensorwaves.data.rng import NumpyUniformRNG, TFUniformRealNumberGenerator
+
+
+class TestNumpyUniformRNG:
+    def test_generate_deterministic(self):
+        rng1 = NumpyUniformRNG(seed=0)
+        rng2 = NumpyUniformRNG(seed=0)
+        assert pytest.approx(rng1(size=3)) == [0.6369617, 0.2697867, 0.0409735]
+        assert pytest.approx(rng2(size=2)) == [0.6369617, 0.2697867]
+        assert pytest.approx(rng2(size=1)) == [0.0409735]
+
+    def test_generate_indeterministic(self):
+        rng1 = NumpyUniformRNG()
+        rng2 = NumpyUniformRNG()
+        with pytest.raises(AssertionError):
+            assert pytest.approx(rng1(size=2)) == rng2(size=2)
+
+    def test_reset_with_seed(self):
+        rng = NumpyUniformRNG(seed=0)
+        assert pytest.approx(rng(size=2)) == [0.6369617, 0.2697867]
+        rng.seed = 0  # reset
+        assert pytest.approx(rng(size=2)) == [0.6369617, 0.2697867]
+
+
+class TestTFUniformRealNumberGenerator:
+    @staticmethod
+    def test_deterministic_call():
+        generator = TFUniformRealNumberGenerator(seed=456)
+        sample = generator(size=3, min_value=-1, max_value=+1)
+        assert pytest.approx(sample) == [-0.38057342, -0.21197986, 0.14724727]
